refactor: separate matchers
