+++ packages/game/src/condition/matcher.ts
@@ -1,4 +1,4 @@
-import { Command } from './command'
+import { Command } from '../command.js'
 
 export interface Matcher {
   match(command: Partial<Command>): boolean
